requirements.txt
appdirs==1.4.4
apturl==0.5.2
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bert-extractive-summarizer==0.5.1
bleach==3.1.1
blinker==1.4
blis==0.4.1
boto3==1.16.13
botocore==1.19.13
branca==0.4.1
Brlapi==0.7.0
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
click-plugins==1.1.1
cligj==0.7.0
colorama==0.4.3
command-not-found==0.3
crcmod==1.7
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
descartes==1.1.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Fiona==1.8.18
flickrapi==2.4.0
folium==0.11.0
future==0.18.2
GDAL==3.0.4
geographiclib==1.50
geojson==2.5.0
geopandas==0.8.1
geopy==2.0.0
gestures==0.2.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.9.0
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==5.6.4
jedi==0.15.2
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libevdev==0.5
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
munch==2.5.0
murmurhash==1.0.4
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.5
neuralcoref==4.0
notebook==6.0.3
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
OpenTimelineIO==0.12.1
osmtogeojson==0.0.2
overpass==0.7
overpy==0.4
OWSLib==0.19.1
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
plotly==4.4.1
preshed==3.0.4
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psycopg2==2.8.4
py4j==0.10.9
pyaaf2==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pykalman==0.9.5
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyspark==3.0.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyudev==0.21.0
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
regex==2020.10.28
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.94
Shapely==1.7.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
spacy==2.3.2
SQLAlchemy==1.3.12
srsly==1.0.3
statsmodels==0.12.0
systemd-python==234
terminado==0.8.2
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.10.1
tokenizers==0.9.2
toml==0.10.1
torch==1.7.0
tornado==5.1.1
tqdm==4.51.0
traitlets==4.3.3
transformers==3.4.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.33
wadllib==1.3.3
wasabi==0.8.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==2.0.0
wikipedia==1.4.0
Wikipedia-API==0.5.4
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
